Added 1e examples for lessons that have not yet been rerecorded
I am working on the Second Edition of this product. As I replace each lesson, I'll add its examples to a folder that has a two-digit chapter number as its name (01, 02, ...) and I'll remove the old version of the examples. The old versions of the examples are in folders with names that begin with "PythonLL1e_lesson"